+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 
 - Chg #63: Raise PHP version to 8.0 (@arogachev)
 - Enh #63: Improve performance (@arogachev)
-- Enh #?: Implement `getByNames()` and `getAccessTree()` methods in `ItemsStorage` (@arogachev)
-- Enh #?: Implement `filterUserItemNames()` method in `AssignmentsStorage` (@arogachev)
-- Chg #?: Rename `$name` argument to `$names` and allow array type for it in `getAllChildren()`, `getAllChildRoles()`,
+- Enh #70: Implement `getByNames()` and `getAccessTree()` methods in `ItemsStorage` (@arogachev)
+- Enh #70: Implement `filterUserItemNames()` method in `AssignmentsStorage` (@arogachev)
+- Chg #70: Rename `$name` argument to `$names` and allow array type for it in `getAllChildren()`, `getAllChildRoles()`,
   `getAllChildPermissions()` methods in `ItemsStorage` (@arogachev)
 
 ## 1.0.0 April 08, 2022